+++ fireapp/lib/main.dart
@@ -2,7 +2,7 @@
 // EXTERNAL
 import 'package:fireapp/global/di.dart';
 import 'package:fireapp/pages/Supervisor/schedulerForm.dart';
-import 'package:fireapp/presentation/constraint_form/constraint_form.dart';
+import 'package:fireapp/presentation/constraint_form/constraint_form_view.dart';
 import 'package:fireapp/presentation/volunteer_information/volunteer_information.dart';
 import 'package:fireapp/presentation/login/login_page.dart';
 import 'package:fireapp/presentation/register/register_page.dart';
