+++ fireapp/lib/presentation/constraint_form/base_input_field.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:intl/intl.dart';
-import 'constraint_form.dart';
+import 'constraint_form_view.dart';
 
 class BaseInputField extends StatelessWidget {
   final TextEditingController controller;
